requirements.txt
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.15.0
rsa==4.7.2
Telethon==1.21.1
